% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{addListElement}
\alias{addListElement}
\title{addListElement}
\usage{
addListElement(theFrame, headers, rows, varName, firstColSize, colSize,
  boldMark, indentStr)
}
\arguments{
\item{theFrame}{List of Lists holding the Lists to be added to, 1st time through just 
contains "term" and "variable", both NULL)}

\item{headers}{Vector of the columns and their titles}

\item{rows}{Vector of text to be processed}

\item{varName}{Name of the variable that created all these rows}

\item{firstColSize}{Length the first (label) column was padded to}

\item{colSize}{Length all other columns were padded to}

\item{boldMark}{Text used to indicate something is bold text.  Ignored if empty}

\item{indentStr}{Text used to indent text.  Ignored if " "}
}
\value{
List of Lists updated with the data from the passed in rows
}
\description{
Make lists for the data frame the output will be added to
}
\author{
m082166
}

