% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{addMethods}
\alias{addMethods}
\title{addMethods}
\usage{
addMethods(results, methods)
}
\arguments{
\item{results}{Vector of strings to add to, will add immediately to end of results}

\item{methods}{List of Methods, where the names are method names, and the values are the 
order the methods appear in the output}
}
\value{
Vector of strings holding Pandoc code to create the table and its methods, if any
}
\description{
Add the methods to the table output
}
\author{
m082166
}

