% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{addModel}
\alias{addModel}
\title{addModel}
\usage{
addModel(rows, useCols, volueCols, modelSum, extra, curRow, colSize, digits,
  pValueDigits, nsmall, nsmall.ratio)
}
\arguments{
\item{rows}{Vector of strings to edit, and possibly add to}

\item{useCols}{Vector of column numbers from modelSum to use, in order to use them}

\item{volueCols}{Vector of 0s and at most 1 "1", specifying if any column is the value col}

\item{modelSum}{The row of the coefficients of modelsum to process, holding the info to add}

\item{extra}{Vector of strings to pad and add to end of the row}

\item{curRow}{Current row to operate on, 0 based}

\item{colSize}{Width to pad each cell to}

\item{digits}{Number of digits to round to when displaying percent or Other data}

\item{pValueDigits}{Number of digits to display for a p-value. Example: 5 ==> in 0.12345}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}

\item{nsmall.ratio}{Minimum number of digits to the right of the decimal point to display 
for the ratio statistics (OR, HR, RR)}
}
\value{
The updated rows
}
\description{
Takes the Vector of the currently filled in rows, as well as rows that have been started but 
not yet completed, and fills in the modelSum info for one more row
}
\author{
m082166
}

