% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{addModelsumTranslations}
\alias{addModelsumTranslations}
\title{addModelsumTranslations}
\usage{
addModelsumTranslations(object, labelTranslations)
}
\arguments{
\item{object}{The data defining the table to display}

\item{labelTranslations}{List where name is the label in the output, and value is the label you 
want displayed e.g. list (q1q3: "Q1, Q3", medsurv = "Median Survival")}
}
\value{
Current translation list
}
\description{
Add all the desired translations from machine produced labels to a human readable ones to the list
}
\author{
m082166
}

