% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{addNumberToEnd}
\alias{addNumberToEnd}
\title{addNumberToEnd}
\usage{
addNumberToEnd(startStr, addNum, size, digits = 5, padChar = " ",
  endText = "")
}
\arguments{
\item{startStr}{Beginning string}

\item{addNum}{Number to put at the end, if NA, or not a number, will just pad to end}

\item{size}{How big to make the string}

\item{digits}{Number of digits to give the number when formatting it, defaults to 5}

\item{padChar}{What to use when padding the string, defaults to ' ', may break if length != 1}

\item{endText}{Text to add after the number, defaults to ""}
}
\value{
String of length size, starting with startStr, ending with addNum
}
\description{
Make a string starting with the starting string, and ending with the passed in number, 
with however many instances of the repeated character are needed to pad out the string to the 
requested length
}
\author{
m082166
}

