% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{addStatistic}
\alias{addStatistic}
\title{addStatistic}
\usage{
addStatistic(rows, stat, startRow, whichStat, colSize, outputTypes, isDate,
  digits, nsmall, pctNSmall, includeTotal)
}
\arguments{
\item{rows}{Vector of strings to edit, and possibly add to}

\item{stat}{The Stats element to process, holding the info to add}

\item{startRow}{Current row to operate on, 0 based}

\item{whichStat}{Which of the items in outputTypes to use}

\item{colSize}{Width to pad each cell to}

\item{outputTypes}{Vector from which to pull output type}

\item{isDate}{If true, show data as dates, if false treat normally}

\item{digits}{Number of digits to round to when displaying percent or Other data}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}

\item{pctNSmall}{Minimum number of digits to the right of the decimal point to display for 
percent numbers}

\item{includeTotal}{TRUE if should include last pre-pValue column, FALSE if shouldn't}
}
\value{
The updated rows
}
\description{
Takes the Vector of the currently filled in rows, as well as rows that have been started but 
not yet completed, and fills in the stats info for one more row
}
\author{
m082166
}

