% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{addToListVector}
\alias{addToListVector}
\title{addToListVector}
\usage{
addToListVector(baseList, title, value)
}
\arguments{
\item{baseList}{List to update}

\item{title}{Name of baseList element to update}

\item{value}{Text to add to the Vector at baseList[[title]]}
}
\value{
baseList after it has been updated
}
\description{
If baseList[[title]] is NULL, make it a Vector holding value.  If it's not null, make a
  vector holding its contents followed by value
}
\author{
m082166
}

