% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{addToRow}
\alias{addToRow}
\title{addToRow}
\usage{
addToRow(rows, curRow, toAdd, sep = " ", padChar = " ")
}
\arguments{
\item{rows}{Vector of strings to edit, and possibly add to}

\item{curRow}{Current row to operate on, 0 based}

\item{toAdd}{Vector of strings to add to the appropriate row}

\item{sep}{Separator to use when pasting together, defaults to " "}

\item{padChar}{Character to use when padding out because len row != len toAdd, defaults to ' '}
}
\value{
The updated rows
}
\description{
A row is defined as a series of non-empty strings, ended by an empty string or by the end of the 
vector.  This method finds the 0 based curRow'th row, and pastes the strings passed in to the 
matching strings in the row.  If there are already more strings in the row than there are to be 
added, the remaining strings are padded with spaces, to the length of the first added string.  
If there are more strings to add than are currently in the row, strings padded to the length of 
the first string in the row are added to the row, then added to<br>
Examples:<br>
Input row: {"foo", "bar"}, toAdd {"baz"} --> {"foo baz", "bar    "}<br>
Input row: {"foo"}, toAdd {"bar", "baz"} --> {"foo bar", "    baz"}<br>
}
\author{
m082166
}

