% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{addTranslations}
\alias{addTranslations}
\title{addTranslations}
\usage{
addTranslations(translations, machine, human)
}
\arguments{
\item{translations}{List to add to where name is the label in the output, and value is the 
label to display, e.g. list (q1q3 = "Q1, Q3", medsurv = "Median Survival")}

\item{machine}{Machine produced labels}

\item{human}{Human readable versions}
}
\value{
Updated translation list
}
\description{
Add translations from machine produced labels to a human readable ones
}
\author{
m082166
}

