% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{addUniqueName}
\alias{addUniqueName}
\title{addUniqueName}
\usage{
addUniqueName(rowNames, newRow)
}
\arguments{
\item{rowNames}{Vector to add to, and test against}

\item{newRow}{New row name to add}
}
\value{
rowNames with newRow concatenated as a unique text string
}
\description{
Add a row name, making it unique if necessary
}
\author{
m082166
}

