% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{arsenal-defunct}
\alias{arsenal-defunct}
\title{Defunct functions in \code{arsenal}}
\description{
Details about defunct functions in \code{arsenal}
}
\details{
\code{length.tableby} was removed in version 2.0.0.

\code{includeNA.character} and \code{includeNA.numeric} were removed in version 2.0.0 and
  replaced with a default method.

\code{rangeTime} was removed in version 1.5.0.
}
\seealso{
\code{\link{arsenal-deprecated}}
}
