% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{arsenal-deprecated}
\alias{arsenal-deprecated}
\title{Deprecated functions in \code{arsenal}}
\description{
Details about deprecated functions in \code{arsenal}
}
\seealso{
\code{\link{arsenal-defunct}}
}
