% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal-package.R
\docType{package}
\name{arsenal-package}
\alias{arsenal}
\alias{arsenal-package}
\title{arsenal: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/arsenal}
  \item Report bugs at \url{https://github.com/jimbrig2011/arsenal/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
