% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{beginsWithPad}
\alias{beginsWithPad}
\title{beginsWithPad}
\usage{
beginsWithPad(testStr)
}
\arguments{
\item{testStr}{String to test}
}
\value{
TRUE if starts with a "pad" / "boundary" character, else FALSE
}
\description{
Reports if a string starts with one of the "boundary characters" that a padded string 
could have been split on, including a space
}
\author{
m082166
}

