% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{capitalizeWords}
\alias{capitalizeWords}
\title{capitalizeWords}
\usage{
capitalizeWords(theStr)
}
\arguments{
\item{theStr}{String to capitalize}
}
\value{
String with each word capitalized
}
\description{
Take a string holding one or more words, and capitalize each word
}
\author{
m082166
}

