% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.control.R
\name{comparison.control}
\alias{comparison.control}
\title{Control settings for `compare` function}
\usage{
comparison.control(tol.num = c("absolute", "percent", "pct"),
  tol.num.val = sqrt(.Machine$double.eps), int.as.num = FALSE,
  tol.char = c("none", "trim", "case", "both"), tol.factor = c("none",
  "levels", "labels"), factor.as.char = FALSE, tol.date = NULL,
  tol.vars = "none", ...)
}
\arguments{
\item{tol.num}{A function, or one of the shortcut character strings; the tolerance to use for numerics.
Shortcut options (can be abbreviated) are:
\code{"absolute"} compare absolute differences,
\code{"percent"} or \code{"pct"} compare percent differences.}

\item{tol.num.val}{Numeric; maximum value of differences allowed in numerics (fed to the function given in \code{tol.num}).}

\item{int.as.num}{Logical; should integers be coerced to numeric before comparison? Default FALSE.}

\item{tol.char}{A function, or one of the shortcut character strings; the tolerance to use for characters.
Shortcut options (can be abbreviated) are
\code{"none"} (compare character strings exactly as they are),
\code{"trim"} (left-justify and trim all trailing white space),
\code{"case"} (allow differences in upper/lower case), and
\code{"both"} (combine \code{"trim"} and \code{"case"}).}

\item{tol.factor}{A function, or one of the shortcut character strings; the tolerance to use for factors.
Shortcut options (can be abbreviated) are
\code{"none"} (match both character labels and numeric levels),
\code{"levels"} (match only the numeric levels), and
\code{"labels"} (match only the labels).}

\item{factor.as.char}{Logical; should factors be coerced to character before comparison? Default FALSE.}

\item{tol.date}{In development...}

\item{tol.vars}{Either \code{"none"} (the default), denoting that variable names are to be matched as-is, or a
character vector denoting equivalence classes for characters in the variable names. See "details", below.}

\item{...}{Other arguments (not in use at this time).}
}
\value{
A list containing the necessary parameters for the \code{\link{compare.data.frame}} function.
}
\description{
Control tolerance definitions for the \code{\link{compare.data.frame}} function.
}
\details{
If not set to \code{"none"} (the default), the \code{tol.vars} argument is a character vector denoting equivalence classes
  for the characters in the variable names. A single character in this vector means to replace that character
  with \code{""}. All other strings in this vector are split by character and replaced by the first character in the string.

E.g., a character vector \code{c("._", "aA", " ")} would denote that the dot and underscore are equivalent (to be translated to a dot),
  that "a" and "A" are equivalent (to be translated to "a"), and that spaces should be removed.

The special character string \code{"case"} in this vector is the same as specifying \code{paste0(letters, LETTERS)}.
}
\examples{
cntl <- comparison.control(
  tol.num = "pct",     # calculate percent differences
  tol.vars = c("case", # ignore case
               "._",   # set all underscores to dots.
               "e")    # remove all letter e's
)
}
\seealso{
\code{\link{compare.data.frame}}, \code{\link{comparison.tolerances}}
}
\author{
Ethan Heinzen
}
