% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.internal.R
\name{comparison.internal}
\alias{comparison.internal}
\alias{n.diffs}
\alias{n.diffs.compare.data.frame}
\alias{n.diffs.summary.compare.data.frame}
\alias{diffs}
\alias{diffs.compare.data.frame}
\alias{diffs.summary.compare.data.frame}
\title{Internal Comparison Functions}
\usage{
n.diffs(object, ...)

\method{n.diffs}{compare.data.frame}(object, ...)

\method{n.diffs}{summary.compare.data.frame}(object, ...)

diffs(object, ...)

\method{diffs}{compare.data.frame}(object, ..., by.var = FALSE)

\method{diffs}{summary.compare.data.frame}(object, ..., by.var = FALSE)
}
\arguments{
\item{object}{An object of class \code{compare.data.frame} or \code{summary.compare.data.frame}.}

\item{...}{Other arguments (not in use at this time).}

\item{by.var}{Logical: should the number of differences by variable be reported, or should
all differences be reported (the default).}
}
\description{
Extract differences and number of differences from a compare object.
}
\seealso{
\code{\link{compare.data.frame}} \code{\link{summary.compare.data.frame}}
}
\author{
Ethan Heinzen
}
