% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{compress}
\alias{compress}
\title{compress}
\usage{
compress(theList)
}
\arguments{
\item{theList}{The list of data to compress}
}
\value{
The Strings concatenated together
}
\description{
Take list of 1+ elements of vectors of equal length, produce a vector with things combined
}
\author{
m082166
}

