% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dates}
\alias{dates}
\alias{start_of_month}
\alias{end_of_month}
\title{Start / End of Month Functions}
\usage{
start_of_month(date)

end_of_month(date)
}
\arguments{
\item{date}{date}
}
\value{
start or end of month as a date
}
\description{
Start / End of Month Functions
}
\examples{
start_of_month(Sys.Date())
end_of_month(Sys.Date())
}
