% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{diff.range}
\alias{diff.range}
\title{Difference Range}
\usage{
\method{diff}{range}(x)
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
returns the difference between min and max value
}
\examples{
x <- rnorm(100)
diff.range(x)

}

