% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.internal.R
\name{diffs}
\alias{diffs}
\alias{n.diff.obs}
\alias{n.diff.obs.compare.data.frame}
\alias{n.diff.obs.summary.compare.data.frame}
\alias{n.diffs}
\alias{n.diffs.compare.data.frame}
\alias{n.diffs.summary.compare.data.frame}
\alias{diffs.compare.data.frame}
\alias{diffs.summary.compare.data.frame}
\title{Extract differences}
\usage{
n.diff.obs(object, ...)

\method{n.diff.obs}{compare.data.frame}(object, ...)

\method{n.diff.obs}{summary.compare.data.frame}(object, ...)

n.diffs(object, ...)

\method{n.diffs}{compare.data.frame}(object, ...)

\method{n.diffs}{summary.compare.data.frame}(object, ...)

diffs(object, ...)

\method{diffs}{compare.data.frame}(object, vars = NULL, ...,
  by.var = FALSE)

\method{diffs}{summary.compare.data.frame}(object, vars = NULL, ...,
  by.var = FALSE)
}
\arguments{
\item{object}{An object of class \code{compare.data.frame} or \code{summary.compare.data.frame}.}

\item{...}{Other arguments (not in use at this time).}

\item{vars}{A character vector of variable names to subset the results to.}

\item{by.var}{Logical: should the number of differences by variable be reported, or should
all differences be reported (the default).}
}
\description{
Extract differences, number of differences, or number of not-shared observations from a \code{compare} object.
}
\seealso{
\code{\link{compare.data.frame}} \code{\link{summary.compare.data.frame}}
}
\author{
Ethan Heinzen
}
