% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{endsWithPad}
\alias{endsWithPad}
\title{endsWithPad}
\usage{
endsWithPad(testStr, testEnd = TRUE)
}
\arguments{
\item{testStr}{String to test}

\item{testEnd}{If TRUE, will test last character, if FALSE will test first character}
}
\value{
TRUE if has a "pad" / "boundary" character in tested place, else FALSE
}
\description{
Reports if a string ends (or starts) with one of the "boundary characters" that a padded string 
could have been split on, including a space
}
\author{
m082166
}

