% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{format.addTranslations}
\alias{format.addTranslations}
\title{format.addTranslations}
\usage{
\method{format}{addTranslations}(object, transList,
  baseList = format.translations, elemCol = "x", nameCol = "name")
}
\arguments{
\item{object}{The data defining the table to display, and its control parameters}

\item{transList}{List where name is the label in the output, and value is the label to display
e.g. list (q1q3 = "Q1, Q3", medsurv = "Median Survival")}

\item{baseList}{List holding any default / starting translations}

\item{elemCol}{The column of object that holds the elements, defaults to "x"}

\item{nameCol}{...?}
}
\value{
Current translation list
}
\description{
Add translations from machine produced labels to a human readable ones
Get translations from object control parameters as well as what was passed in
}
\author{
m082166
}

