% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{format.other}
\alias{format.other}
\title{format.other}
\usage{
\method{format}{other}(data, colSize, outputTypes, whichStat, isDate, digits,
  nsmall, pctNSmall)
}
\arguments{
\item{data}{Data to display, be it number or date}

\item{colSize}{Width of the output string}

\item{outputTypes}{Vector from which to pull output type}

\item{whichStat}{Which of the items in outputTypes to use}

\item{isDate}{If true, show data as dates, if false treat normally}

\item{digits}{Number of digits to round to when displaying percent or Other data}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}

\item{pctNSmall}{Minimum number of digits to the right of the decimal point to display for 
percent numbers}
}
\value{
String of length colSize
}
\description{
Format data output, according to the output type specified, padded to fill to colSize
}
\author{
m082166
}

