% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{format.three}
\alias{format.three}
\title{format.three}
\usage{
\method{format}{three}(first, second, third, colSize, outputTypes, whichStat,
  isDate, digits, nsmall)
}
\arguments{
\item{first}{First number to display}

\item{second}{Second number to display}

\item{third}{Third number to display}

\item{colSize}{Width of the output string}

\item{outputTypes}{Vector from which to pull output type}

\item{whichStat}{Which of the items in outputTypes to use}

\item{isDate}{If true, show data as dates, if false treat normally}

\item{digits}{Number of digits to round to when displaying percent or Other data}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}
}
\value{
String of length colSize
}
\description{
Format three number output, according to the output type specified, padded to fill to colSize
}
\author{
m082166
}

