% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{formatElement}
\alias{formatElement}
\title{formatElement}
\usage{
formatElement(element, lineSize, firstColSize, colSize, includeTotal, hasPValue,
  translations, digits, pValueDigits, nsmall, pctNSmall, boldMark, indentStr,
  collapse, methods)
}
\arguments{
\item{element}{List to get information from, whose first item must be the statistics}

\item{lineSize}{Length each non-blank line should be padded to}

\item{firstColSize}{Length the first (label) column should be padded to}

\item{colSize}{Length all other columns should be padded to}

\item{includeTotal}{TRUE if should include last pre-pValue column, FALSE if shouldn't}

\item{hasPValue}{TRUE if has column for p-values, FALSE if shouldn't}

\item{translations}{The List to use for conversion of labels}

\item{digits}{Maximum number of digits to display for floating point numbers}

\item{pValueDigits}{Number of digits to display for a p-value. Example: 5 ==> in 0.12345}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}

\item{pctNSmall}{Minimum number of digits to the right of the decimal point to display for 
percent numbers.  If NULL, use 'nsmall', if that is NULL use 'digits' to determine everything}

\item{boldMark}{String to use to mark text as bold}

\item{indentStr}{String to use to indent something one space}

\item{collapse}{If true, data might have been collapsed, and needs to be tested}

\item{methods}{List of methods and when they were first seen, or NULL if not tracking that}
}
\value{
List of two elements:
strings: Vector of strings holding Pandoc code to create a row in a table representing element
methods: Updated list of methods, or NULL if methods was NULL
}
\description{
Return a List with two elements: 
The vector holding the lines of a row in the table, defined by element, in Pandoc format
Updated list of methods used by this tableby object
}
\author{
m082166
}

