% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{formatModelSum}
\alias{formatModelSum}
\title{formatModelSum}
\usage{
formatModelSum(element, lineSize, firstColSize, colSize, translations, digits,
  pValueDigits, nsmall, nsmall.ratio, boldMark, showIntercept, showAdjust,
  showCols, showOnce, fieldName, oneTimeFieldName, pFootnote)
}
\arguments{
\item{element}{List to get information from, whose fieldName item must be the statistics}

\item{lineSize}{Length each non-blank line should be padded to}

\item{firstColSize}{Length the first (label) column should be padded to}

\item{colSize}{Length all other columns should be padded to}

\item{translations}{The List to use for conversion of labels}

\item{digits}{Maximum number of digits to display for floating point numbers}

\item{pValueDigits}{Number of digits to display for a p-value. Example: 5 ==> in 0.12345}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If NULL, use 'digits' to determine everything}

\item{nsmall.ratio}{Minimum number of digits to the right of the decimal point to display 
for the ratio statistics (OR, HR, RR)}

\item{boldMark}{String to use to mark text as bold}

\item{showIntercept}{TRUE if should show the Intercept for each line, FALSE if shouldn't}

\item{showAdjust}{If TRUE show all rows, if false hide the "adjust" rows}

\item{showCols}{List of items that will be output as columns from each row, in order}

\item{showOnce}{List of items that will be output as columns only on 1st row, in order}

\item{fieldName}{The name of the List element that holds the matrix with the headers}

\item{oneTimeFieldName}{...?}

\item{pFootnote}{If TRUE add a footnote describing the test used to generate the p value}
}
\value{
Vector of strings holding Pandoc code to create a row in a table representing element
}
\description{
Return a List with two elements: 
The vector holding the lines of a row in the table, defined by element, in Pandoc format
Updated list of methods used by this modelsum object
}
\author{
m082166
}

