% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.internal.R, R/internal.functions.R
\name{freqlist.internal}
\alias{freqlist.internal}
\alias{is.freqlist}
\alias{is.summary.freqlist}
\alias{[.freqlist}
\alias{labels.freqlist}
\alias{labels<-.freqlist}
\alias{merge.freqlist}
\title{Helper functions for freqlist}
\usage{
is.freqlist(x)

is.summary.freqlist(x)

\method{[}{freqlist}(x, i, j, ...)

\method{labels}{freqlist}(object, ...)

\method{labels}{freqlist}(x) <- value

\method{merge}{freqlist}(x, y, all = TRUE, ...)
}
\arguments{
\item{x, y}{A \code{freqlist} object.}

\item{i}{A vector to index \code{x} with: either names of variables, a numeric vector, or a logical vector of appropriate length.
\code{i} indexes the x-variables, and \code{j} indexes the by-variables.}

\item{j}{A vector to index \code{x} with: either names of variables, a numeric vector, or a logical vector of appropriate length.
\code{i} indexes the x-variables, and \code{j} indexes the by-variables.}

\item{...}{Other arguments, or a vector of indices for extracting.}

\item{object}{An object}

\item{value}{A list of new labels.}

\item{all}{Logicals, denoting which terms to keep if not all are in common.}
}
\description{
A set of helper functions for \code{\link{freqlist}}.
}
