% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{geom_mark_outlier}
\alias{geom_mark_outlier}
\title{get outlir}
\usage{
geom_mark_outlier(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
get outlier
}
\examples{



}

