% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{geom_name_outlier}
\alias{geom_name_outlier}
\title{get outlir}
\usage{
geom_name_outlier(mapping = NULL, data = NULL,
  position = position_stack(1.08), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
get outlier
}
\examples{



}

