% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-outlier.R
\name{get_outlier}
\alias{get_outlier}
\title{get outlier}
\usage{
get_outlier(x, y = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector or NULL}
}
\value{
Numeric Index
}
\description{
get outlier
}
\examples{
get_outlier(mtcars$mpg, mtcars$hp)
mtcars[get_outlier(mtcars$mpg, mtcars$hp), ]

}

