% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{getColsToShow}
\alias{getColsToShow}
\title{getColsToShow}
\usage{
getColsToShow(control, family, element, fieldName, oneTimeFieldName,
  translations)
}
\arguments{
\item{control}{Control object, holding the columns we're displaying}

\item{family}{Stats family of object we're displaying}

\item{element}{1st element we're displaying, to get column names / locations}

\item{fieldName}{The name of the List element for the "show every row" data}

\item{oneTimeFieldName}{The name of the List element for the "show once" data}

\item{translations}{The List to use for conversion of labels}
}
\value{
List of Vectors of names of rows to output
showCols: fields to be pulled from every row of element$coeff
showOnce: fields to be pulled from element$glance, one time only
}
\description{
Takes the Vector of the currently filled in rows, as well as rows that have been started but 
not yet completed, and fills in the modelSum info for one more row
}
\author{
m082166
}

