% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{getColsToUse}
\alias{getColsToUse}
\title{getColsToUse}
\usage{
getColsToUse(showCols, colTitles)
}
\arguments{
\item{showCols}{List of items that will be output as columns from each row, in order}

\item{colTitles}{List of column titles for modelSum, so can determine which to display}
}
\value{
Columns to be usedm in order
}
\description{
Generate a Vector holding the numbers of the columns to be used, in the order to be used
}
\author{
m082166
}

