% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{getMatrixNames}
\alias{getMatrixNames}
\title{getMatrixNames}
\usage{
getMatrixNames(element, fieldName, doRow = FALSE, translations = NULL)
}
\arguments{
\item{element}{A List of Lists}

\item{fieldName}{The name of the List element of interest}

\item{doRow}{If TRUE get row names, if FALSE get column names, defaults to FALSE}

\item{translations}{The List to use for conversion of labels, so can use the proper name length}
}
\value{
Vector of Strings, the row or column headers of the matrix
}
\description{
Get the row or column names for the matrix in fieldName
}
\author{
m082166
}

