% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{getName}
\alias{getName}
\title{getName}
\usage{
getName(nameText, boldMark, indentStr)
}
\arguments{
\item{nameText}{Text to parse}

\item{boldMark}{Text used to indicate something is bold text.  Ignored if empty}

\item{indentStr}{Text used to indent text.  Ignored if " "}
}
\value{
String holding the ceaned up text.  Will clean either boldMark or indentStr
}
\description{
Extract the row name from the text
}
\author{
m082166
}

