% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{getPct}
\alias{getPct}
\title{getPct}
\usage{
getPct(doPct)
}
\arguments{
\item{doPct}{Logical. If \code{TRUE}, returns "\%" else "".}
}
\value{
The appropriate string: "\%" or ""
}
\description{
Return the proper "pct" string for the specified stat. "\%" if it's percent data, "" if not
}
\author{
m082166
}

