% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{getRowInfo}
\alias{getRowInfo}
\title{getRowInfo}
\usage{
getRowInfo(object, showIntercept, showAdjust, yCol, translations)
}
\arguments{
\item{object}{The data defining the table to display}

\item{showIntercept}{TRUE if included the Intercept with each variable}

\item{showAdjust}{TRUE if included the adjust variable results for each variable}

\item{yCol}{Name of the column in glance that holds the y column value}

\item{translations}{Translations}
}
\value{
List with one element per model of object, holding # rows & y variable
}
\description{
For each model, get the number of rows that are part of that model, and the y variable associated 
with it
}
\author{
m082166
}

