% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{getStartingLine}
\alias{getStartingLine}
\title{getStartingLine}
\usage{
getStartingLine(rows, curRow)
}
\arguments{
\item{rows}{Vector of strings to look through}

\item{curRow}{Current row to operate on, 0 based}
}
\value{
The location in rows (1 based) holding the beginning of "row" curRow, 
or the length of rows if rows doesn't hold that many "rows"
}
\description{
A row is defined as a series of non-empty strings, ended by an empty string or by the end of the 
vector.  This method finds the 0 based curRow'th row, and returns its location within rows
}
\author{
m082166
}

