% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.in.R
\name{\%nin\%}
\alias{\%nin\%}
\alias{nin}
\title{Not in}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}
}
\value{
The negation of \code{\link{\%nin\%}}.
}
\description{
The not-in operator for R.
}
\examples{
1 \%nin\% 2:10
c("a", "b") \%nin\% c("a", "c", "d")
}
\author{
Raymond Moore
}
\seealso{
\code{\link{\%in\%}}, \code{\link[Hmisc]{\%nin\%}}
}

