% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/notin.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Description}
\usage{
x \%not in\% table

x \%not in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}

\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
None

None
}
\description{
match returns a vector of non matches

match returns a vector of non matches
}
\examples{
"Hans" \%notin\% c("Fritz","Rolf", "Reto")

"Hans" \%notin\% c("Fritz","Rolf", "Reto")

}

