% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notlike.R
\name{\%notlike\%}
\alias{\%notlike\%}
\title{Description}
\usage{
x \%notlike\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
None
}
\description{
match returns a vector of non matches
}
\examples{
"Hans" \%notin\% c("Fritz","Rolf", "Reto")

}

