% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{integerDigits}
\alias{integerDigits}
\title{integerDigits}
\usage{
integerDigits(aNumber)
}
\arguments{
\item{aNumber}{The number to do the calculations for}
}
\value{
An integer from 0 up, 0 if input isn't a number
}
\description{
Compute the number of integer digits (i.e. significant digits to the left of the decimal place) 
a positive number has
}
\author{
m082166
}

