% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{internal.functions}
\alias{internal.functions}
\alias{smart.split}
\alias{replace2}
\title{Internal Functions}
\usage{
smart.split(string, width = Inf, min.split = -Inf)

replace2(x, list, values)
}
\arguments{
\item{string}{A character vector}

\item{width}{Either \code{Inf} or \code{NULL} to specify no splitting,
or a positive integer giving the largest allowed string length.}

\item{min.split}{Either \code{-Inf} or \code{NULL} to specify no
lower bound on the string length, or a positive integer giving the minimum string length.}

\item{x}{vector}

\item{list}{an index vector}

\item{values}{replacement values}
}
\value{
For \code{smart.split}, a list of the same length as \code{string}, with each element being
  the "intelligently" split string.

  For \code{replace2}, a vector with the proper values replaced.
}
\description{
Internal Functions
}
\seealso{
\code{\link[base]{replace}}
}
