% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isFloatOut}
\alias{isFloatOut}
\title{isFloatOut}
\usage{
isFloatOut(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if floating point output, else FALSE
}
\description{
Returns TRUE if this stat's data should be displayed as floating point number(s)
}
\author{
m082166
}

