% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isListOut}
\alias{isListOut}
\title{isListOut}
\usage{
isListOut(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if list output, else FALSE
}
\description{
Returns TRUE if this stat's data should be displayed as a list A[, B]*
}
\author{
m082166
}

