% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isMedListOut}
\alias{isMedListOut}
\title{isMedListOut}
\usage{
isMedListOut(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if median list output, else FALSE
}
\description{
Returns TRUE if this stat's data should be displayed as a median list A (B, C)
}
\author{
m082166
}

