% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isMedRange}
\alias{isMedRange}
\title{isMedRange}
\usage{
isMedRange(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if median range output, else FALSE
}
\description{
Returns TRUE if this stat's two valued data should be displayed as a median range A (B - C)
}
\author{
m082166
}

