% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isOther}
\alias{isOther}
\title{isOther}
\usage{
isOther(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return TRUE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if other, else FALSE
}
\description{
Returns TRUE if this stat's two valued data should be displayed as an "other" type A (B)
}
\author{
m082166
}

