% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{isPct}
\alias{isPct}
\title{isPct}
\usage{
isPct(outputTypes, whichStat)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}
}
\value{
TRUE if percent output, else FALSE
}
\description{
Returns TRUE if this stat's two valued data should be displayed as a count / percent A (B%)
}
\author{
m082166
}

