% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftright.R
\name{leftright}
\alias{leftright}
\alias{left}
\alias{right}
\title{Left Right}
\usage{
left(string, num_chars)

right(string, num_chars)
}
\arguments{
\item{string}{Character string to pull from.}

\item{num_chars}{Number of characters to pull.}
}
\value{
Adjusted character string.
}
\description{
R functions mirroring Excel's left() and right() functions.
}
\examples{
left("leftright", 4)
right("leftright", 5)
}
