% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files}
\alias{list_files}
\title{List files removing all temporary / internal unwanted files}
\usage{
list_files(path, pat = NULL, full = TRUE, named = FALSE)
}
\arguments{
\item{path}{Path to list files from}

\item{pat}{Optional pattern to match. Defaults to NULL}

\item{full}{Logical whether to return full path or not. Defaults to TRUE}

\item{named}{Logical to return names or not}
}
\value{
A character vector containing the names of the files in the specified
  directories excluding any temporary unwanted files (i.e. "~$file.xlsx").
}
\description{
List files removing all temporary / internal unwanted files
}
\examples{
list_files(getwd())
}
