% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{logrank}
\alias{logrank}
\title{logrank}
\usage{
logrank(x, x.by)
}
\arguments{
\item{x}{surv variable}

\item{x.by}{by, categorical variable}
}
\value{
test output with $method and $p.value
}
\description{
survdiff logrank test
}

