% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{lookupHumanTitle}
\alias{lookupHumanTitle}
\title{lookupHumanTitle}
\usage{
lookupHumanTitle(label, translations = format.translations)
}
\arguments{
\item{label}{The label to convert, or a vector of labels to convert}

\item{translations}{The List to use for conversion of labels, defaults to format.translations}
}
\value{
More human readable version of the label, if have one, else the passed in label.
If was passed a Vector, will return a Vector of translations
}
\description{
Take a string, and see if we have a human readable version of that string
}
\author{
m082166
}

