% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeCellHeader}
\alias{makeCellHeader}
\title{makeCellHeader}
\usage{
makeCellHeader(text, size, leftJustify = FALSE, rightJustify = FALSE)
}
\arguments{
\item{text}{The text of the header}

\item{size}{The size the header must be padded to}

\item{leftJustify}{If TRUE, will left justify each column, defaults to FALSE}

\item{rightJustify}{If TRUE, will right justify each column, defaults to FALSE 
When both leftJustify and rightJustify are FALSE, columns are centered}
}
\value{
String of length size
}
\description{
Create a string holding the complete header for a cell, of length size (which must be >= length 
of text).  Defaults to center justified text , which requires that size >= length of text + 2
Input is not currently validated, caller responsible for setting values correctly
}
\author{
m082166
}

