% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeCenteredStr}
\alias{makeCenteredStr}
\title{makeCenteredStr}
\usage{
makeCenteredStr(startStr, size, padChar = " ", lineSplit = "\\n",
  sizeLimit = FALSE)
}
\arguments{
\item{startStr}{String providing the text to get centered}

\item{size}{Size to pad the string out to}

\item{padChar}{Character to pad out string to length size, defaults to ' ', 
will cause problems if not length 1}

\item{lineSplit}{String to split startStr into separate lines, defaults to "\\n"}

\item{sizeLimit}{If TRUE, will never output a line length > size, if FALSE will try to avoid 
over-long lines, but a "word" longer than size will just be its own line}
}
\value{
String of length size, or Vector of strings of length size
}
\description{
Make a string consisting of the starting string, surrounded by however many instances of padChar 
are needed to pad out the string to the requested length.
}
\details{
If nchar (startStr) > size, will split startStr at whitespace so smaller than size, 
and will then return a Vector of centered strings of decreasing length.<br>
If there are lineSplit strings in startStr, will first be split into separate lines, and 
each line will be centered separately
}
\author{
m082166
}

