% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{makeCountHeader}
\alias{makeCountHeader}
\title{Make a string of the form "name (N=count)"}
\usage{
makeCountHeader(name, count)
}
\arguments{
\item{name}{The label that's getting a count (i.e. "High")}

\item{count}{The count that goes with this label}
}
\value{
String of the form "name (N=count)"
}
\description{
Make a string of the form "name (N=count)"
}
\author{
m082166
}

