% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeDashStr}
\alias{makeDashStr}
\title{makeDashStr}
\usage{
makeDashStr(count, theChar = "-")
}
\arguments{
\item{count}{Size of the returned string}

\item{theChar}{Character to use when building the string, defaults to '-'}
}
\value{
String of length count
}
\description{
Make a string consisting of count instances of the repeated character
}
\author{
m082166
}

