% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{makeDataFrame}
\alias{makeDataFrame}
\title{makeDataFrame}
\usage{
makeDataFrame(headers, frameLists)
}
\arguments{
\item{headers}{Vector of most of the columns and their titles}

\item{frameLists}{List of lists holding the data for the data frame}
}
\value{
Data Frame with all the elements set up and filled in
}
\description{
Make the tableby data frame and add the output to it
}
\author{
m082166
}

