% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{makeDF}
\alias{makeDF}
\title{makeDF}
\usage{
makeDF(baseList, rowNames, nonNumeric)
}
\arguments{
\item{baseList}{List holding the contents of the data.frame to be}

\item{rowNames}{Names of the rows, will be added as column "term"}

\item{nonNumeric}{Names of non-numeric columns of the resulting data.frame}
}
\value{
A data.frame holding the modelsum, with every column not in nonNumeric made numeric
}
\description{
Create a data.frame, updating the "term" column to have the rowNames as its elements, and 
setting its columns to be numeric or not as appropriate
}
\author{
m082166
}

