% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeHeader}
\alias{makeHeader}
\title{makeHeader}
\usage{
makeHeader(group, minColSize, includeTotal, hasPValue, pValueTitle,
  labelSize = 1.2, leftJustify = FALSE, rightJustify = FALSE)
}
\arguments{
\item{group}{Data we're making the table about}

\item{minColSize}{Minimum size of the first column (which will hold label info for a row)}

\item{includeTotal}{TRUE if should include last pre-pValue column, FALSE if shouldn't}

\item{hasPValue}{TRUE if should have column for p-values, FALSE if shouldn't}

\item{pValueTitle}{Title for pValue, only matters if hasPValue is TRUE}

\item{labelSize}{Relative size difference between label column and other columns. 
Default is 1.2: label column ~20\% bigger than other columns}

\item{leftJustify}{If TRUE, will left justify each column, defaults to FALSE}

\item{rightJustify}{If TRUE, will right justify each column, defaults to FALSE 
When both leftJustify and rightJustify are FALSE, columns are centered}
}
\value{
List holding the lines of the header defined by group plus
lineSize: the length of a full line, 
firstColSize: the length of the first column, 
colSize: the length of each other column, 
header: The lines of the header
The last element is the last line of the output, to go after the body of the output
headers: Vector of the plain header titles for each column
}
\description{
Make the Pandoc format header for the table specified by group
}
\author{
m082166
}

