% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeHeaders}
\alias{makeHeaders}
\title{makeHeaders}
\usage{
makeHeaders(group, includeTotal, hasPValue, pValueTitle)
}
\arguments{
\item{group}{The data that will be turned into a table}

\item{includeTotal}{TRUE if should include last pre-pValue column, FALSE if shouldn't}

\item{hasPValue}{TRUE if should have column for p-values, FALSE if shouldn't}

\item{pValueTitle}{Title for pValue, only matters if hasPValue is TRUE}
}
\value{
A Vector of the column headers, given the data in group, 
skipping the first (blank, label) header
}
\description{
Make the unpadded header for each column other than the label column
}
\author{
m082166
}

