% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeIndentedStr}
\alias{makeIndentedStr}
\title{makeIndentedStr}
\usage{
makeIndentedStr(startStr, size, padChar = " ", indent = 3,
  indentStr = "&nbsp;")
}
\arguments{
\item{startStr}{The text that will be indented and displayed}

\item{size}{Padded width of resulting strings}

\item{padChar}{Character to pad out strings to length size, defaults to ' ', 
will cause problems if not nchar 1}

\item{indent}{Number of spaces to indent startStr, defaults to 3}

\item{indentStr}{String to use to indent something one space, defaults to "&nbsp;" 
(HTML non-breaking space)}
}
\value{
Vector of one or more strings holding the Pandoc code required to display the indented 
string
}
\description{
Make a string consisting of indent copies of indentStr followed by the starting string, 
all this broken across as many lines as need to so each line is length size or less, 
each line followed by however many instances of the repeated character, which defaults to ' ', 
are needed to pad out each line to the requested length
}
\author{
m082166
}

