% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeLimitedNumber}
\alias{makeLimitedNumber}
\title{makeLimitedNumber}
\usage{
makeLimitedNumber(addNum, digits)
}
\arguments{
\item{addNum}{Number to use}

\item{digits}{Number of digits to give the number when formatting it}
}
\value{
String holding addNum, or "< 10^-digits"
}
\description{
Make a string with a number, or "< x", where X is minimum number to show
}
\author{
m082166
}

