% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{makeModelSumHeader}
\alias{makeModelSumHeader}
\title{makeModelSumHeader}
\usage{
makeModelSumHeader(showCols, showOnce, minColSize, labelSize = 1.2,
  translations, leftJustify = FALSE, rightJustify = FALSE)
}
\arguments{
\item{showCols}{List of items that will be output as columns from each row, in order}

\item{showOnce}{List of items that will be output as columns only on 1st row, in order}

\item{minColSize}{Minimum size of the first column (which will hold label info for a row)}

\item{labelSize}{Relative size difference between label column and other columns. 
Default is 1.2: label column ~20\% bigger than other columns}

\item{translations}{The List to use for conversion of labels}

\item{leftJustify}{If TRUE, will left justify each column, defaults to FALSE}

\item{rightJustify}{If TRUE, will right justify each column, defaults to FALSE 
When both leftJustify and rightJustify are FALSE, columns are centered}
}
\value{
List holding the lines of the header defined by group plus
lineSize: The length of a full line, 
firstColSize: The length of the first column, 
colSize: The length of each other column, 
header: The lines of the header
The last element is the last line of the output, to go after the body of the output
}
\description{
Make the Pandoc format header for the table
}
\author{
m082166
}

