% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{makeModelSumHeaders}
\alias{makeModelSumHeaders}
\title{makeModelSumHeaders}
\usage{
makeModelSumHeaders(object, showCols, showOnce, translations)
}
\arguments{
\item{object}{The data that will be turned into a table}

\item{showCols}{List of items that will be output as columns from each row, in order}

\item{showOnce}{List of items that will be output as columns only on 1st row, in order}

\item{translations}{The List to use for conversion of labels}
}
\value{
A Vector of the column headers, given the data in object and extras, 
skipping the first (blank, label) header
}
\description{
Make the unpadded header for each column other than the label column
}
\author{
m082166
}

