% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makePaddedStr}
\alias{makePaddedStr}
\title{makePaddedStr}
\usage{
makePaddedStr(startStr, size, padChar = " ")
}
\arguments{
\item{startStr}{String providing the text to get padded out}

\item{size}{Size to pad the string out to}

\item{padChar}{Character to pad out string to length size, defaults to ' ', 
will cause problems if not length 1}
}
\value{
String of length size, or Vector of strings of length size, if nchar (startStr) > size
}
\description{
Make a string consisting of the starting string, followed by however many instances of the 
repeated character are needed to pad out the string to the requested length.
}
\details{
If nchar (startStr) > size, will try to split startStr at reasonable points so smaller than size, 
if that fails will simply break it at size length, & will then return a Vector of padded strings
}
\author{
m082166
}

