% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{makeReference}
\alias{makeReference}
\title{makeReference}
\usage{
makeReference(methods, method)
}
\arguments{
\item{methods}{List of methods and when they were first seen, or NULL if not tracking that}

\item{method}{Method used this time}
}
\value{
List of two elements:
ref: Text specifying the superscript for the passed in method
methods: Updated list of methods, or NULL if methods was NULL
}
\description{
Return a list with two elements: 
The text specifying the superscript referencing the method used this time
Updated list of methods used by this tableby object
}
\author{
m082166
}

