% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tableby.R
\name{makeSummary.tableby}
\alias{makeSummary.tableby}
\title{Format the information in object as a Table using Pandoc coding or plain text<br/>
If doText is TRUE, cat it to stdout, else just return the data.frame}
\usage{
makeSummary.tableby(doText, object, title, labelTranslations, digits, nsmall,
  nsmall.pct, digits.test, text, removeBlanks, labelSize, test, test.pname,
  pfootnote, total)
}
\arguments{
\item{doText}{Do we print text output, or only do the data.frame}

\item{object}{An object of class \code{tableby}, made by the \code{\link{tableby}} function.}

\item{title}{Title that will appear on the top of the header in the pretty-table rendering 
of the tableby object}

\item{labelTranslations}{All labels that are to appear in the pretty rendering of the \code{tableby} 
results have both summary-statistic labels that are replaced by a formal label 
(e.g., \code{meansd} by \code{"Mean (SD)"}), and the variables from the formula can be replaced 
by a more formal name.}

\item{digits}{Digits to round for significant digits of numeric, non-integer values. 
If \code{digits.test} is not set, \code{digits} is used for that setting.}

\item{nsmall}{Minimum number of digits to the right of the decimal point to display 
for floating point numbers.  If \code{NA} (default), it uses the value from 
\code{object$control$nsmall}. Allowed non-\code{NA} values are \code{0 <= nsmall <= 20}.}

\item{nsmall.pct}{Minimum number of digits to the right of the decimal point to display 
for percent numbers.  If \code{NA} (default), it uses the value from \code{object$control$nsmall.pct}.}

\item{digits.test}{Significant digits by which to round for numeric test statistic p-values, 
if the test was performed.}

\item{text}{Logical, tell R to print the raw text version of the summary to the screen. 
Default is \code{FALSE}, but recommended to be \code{TRUE} for interactive R session development.}

\item{removeBlanks}{Logical, remove extra blanks in the pretty rendering of the table}

\item{labelSize}{Relative size difference between label column and other columns. 
Default is 1.2: label column ~20\% bigger than other columns.}

\item{test}{Logical, denoting whether the "p value" value should be printed. 
If \code{NA} (default), it uses the value from \code{object$control$test}.}

\item{test.pname}{Title for p-value (only matters if test is \code{TRUE}; default is "p value").}

\item{pfootnote}{Logical, denoting whether to add a footnote describing the test used to 
generate the p value. Default is \code{FALSE}.}

\item{total}{Logical, denoting whether to include the "total" value. 
If \code{NA} (default), it uses the value from \code{object$control$total}.}
}
\value{
Results are cat'ed to stdout, and returned invisibly as a Vector of Strings
}
\description{
Format the information in object as a Table using Pandoc coding or plain text<br/>
If doText is TRUE, cat it to stdout, else just return the data.frame
}
\author{
m082166
}

