% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{makeTitleCell}
\alias{makeTitleCell}
\title{makeTitleCell}
\usage{
makeTitleCell(element, colSize, translations, boldMark, indentStr, collapse)
}
\arguments{
\item{element}{List to get information from, whose first item must be the statistics}

\item{colSize}{Width to pad the output to}

\item{translations}{The List to use for conversion of labels}

\item{boldMark}{String to use to mark something as bold}

\item{indentStr}{String to use to indent something one space}

\item{collapse}{...?}
}
\value{
Vector holding the strings necessary to represent the rows of element, 
each row separated by a blank string
}
\description{
Return an array of the lines needed to make the label cell, given the data in element, 
taking into account the maximum allowed width specified by colSize, 
which must be >= 4 + length of the name of element
}
\author{
m082166
}

