% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{maxElementNameLen}
\alias{maxElementNameLen}
\title{maxElementNameLen}
\usage{
maxElementNameLen(elements, translations)
}
\arguments{
\item{elements}{A List of Lists}

\item{translations}{The List to use for conversion of labels, so can use the proper name length}
}
\value{
The nchar length of the longest name from element's sub-lists, 
as translated via translations
}
\description{
Return the length of the longest string among the names of elements
}
\author{
m082166
}

