% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{maxMatrixNameLen}
\alias{maxMatrixNameLen}
\title{maxMatrixNameLen}
\usage{
maxMatrixNameLen(elements, translations, fieldName, doRow = FALSE)
}
\arguments{
\item{elements}{A List of Lists}

\item{translations}{The List to use for conversion of labels, so can use the proper name length}

\item{fieldName}{The name of the List element of interest}

\item{doRow}{If TRUE get row names, if FALSE get column names, defaults to FALSE}
}
\value{
The nchar length of the longest name from element's sub-lists, 
as translated via translations
}
\description{
Return the length of the longest string among the names of elements
}
\author{
m082166
}

