% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{maxStrLen}
\alias{maxStrLen}
\title{maxStrLen}
\usage{
maxStrLen(strings)
}
\arguments{
\item{strings}{Vector of Strings}
}
\value{
The nchar length of the longest string in the Vector
}
\description{
Return the length of the longest string in a Vector of strings
}
\author{
m082166
}

