% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{medSurv}
\alias{medSurv}
\title{medSurv}
\usage{
medSurv(x, ...)
}
\arguments{
\item{x}{kaplan-meier summary object, used within tableby}

\item{...}{other arguments}
}
\value{
vector of median subjects who have survived by time point
}
\description{
Calculate median survival
}

