% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{minStrLen}
\alias{minStrLen}
\title{minStrLen}
\usage{
minStrLen(strings)
}
\arguments{
\item{strings}{Vector of Strings}
}
\value{
The nchar length of the shortest string in the Vector
}
\description{
Return the length of the shortest string in a Vector of strings
}
\author{
m082166
}

