% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockstudy.R
\docType{data}
\name{mockstudy}
\alias{mockstudy}
\alias{muck_up_mockstudy}
\title{Mock study data for examples}
\format{
A data frame with 1499 observations on the following 15 variables:
  \describe{
    \item{\code{case}}{a numeric identifier-patient ID}
    \item{\code{age}}{age in years}
    \item{\code{arm}}{treatment arm divided into 3 groups, character string }
    \item{\code{sex}}{a factor with levels \code{Male} \code{Female}}
    \item{\code{race}}{self-reported race/ethnicity, character string}
    \item{\code{fu.time}}{survival or censoring time in years}
    \item{\code{fu.stat}}{censoring status; 1=censor, 2=death}
    \item{\code{ps}}{integer, ECOG performance score }
    \item{\code{hgb}}{numeric, hemoglobin count}
    \item{\code{bmi}}{numeric, body mass index, kg/m^2}
    \item{\code{alk.phos}}{numeric, alkaline phosphatase}
    \item{\code{ast}}{numeric, aspartate transaminase }
    \item{\code{mdquality.s}}{integer, LASA QOL 0=Clinically Deficient, 1=Not Clinically Deficient  }
    \item{\code{age.ord}}{an ordered factor split of age, with levels
      \code{10-19} < \code{20-29} < \code{30-39} < \code{40-49} <
        \code{50-59} < \code{60-69} < \code{70-79} < \code{80-89}}
  }

An object of class \code{data.frame} with 1499 rows and 14 columns.
}
\usage{
mockstudy

muck_up_mockstudy()
}
\description{
Mock clinical study data for examples to test data manipulation and statistical functions.
  The function \code{muck_up_mockstudy()} is used in examples for \code{\link{comparedf}}.
}
\examples{
data(mockstudy)
str(mockstudy)
}
\keyword{datasets}
