% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsum.families.R
\name{modelsum.family}
\alias{modelsum.family}
\alias{survival}
\alias{ordinal}
\title{Family functions for modelsum}
\usage{
survival()

ordinal(method = c("logistic", "probit", "loglog", "cloglog", "cauchit"))
}
\arguments{
\item{method}{See \code{MASS::\link[MASS]{polr}}.}
}
\value{
A list, in particular with element \code{family}.
}
\description{
A set of family functions for \code{\link{modelsum}}.
}
\seealso{
\code{\link[stats]{family}}, \code{\link[survival]{coxph}}, \code{\link[MASS]{polr}}
}
