% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{myFormat}
\alias{myFormat}
\title{myFormat}
\usage{
myFormat(number, digits, nsmall, isDate = FALSE, doTrim = TRUE)
}
\arguments{
\item{number}{String to format}

\item{digits}{Number of digits to display for number.  Will be passed to format.
If nsmall is NULL, will set to max(0, digits - number of integer digits in number)}

\item{nsmall}{Minimum number of non-zero digits to the right of the decimal point to display.
Will trim off any ending 0s after decimal place}

\item{isDate}{Is it a Date?  If TRUE, return it as.character}

\item{doTrim}{If doTrim is false, won't trim, otherwise will.  Default is TRUE}
}
\value{
Resulting String, will not be empty unless number was empty
}
\description{
Format a number, adjusting nsmall as appropriate, and remove all trailing 0s, and a trailing "."
}
\author{
m082166
}

