% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{myStrJoin}
\alias{myStrJoin}
\title{myStrJoin}
\usage{
myStrJoin(theStr, firstColSize, colSize)
}
\arguments{
\item{theStr}{String, or vector of strings, to join}

\item{firstColSize}{The length of the first column}

\item{colSize}{The length of each other column}
}
\value{
Contents of each column de-Pandoced and formed into a single un-padded string
}
\description{
Take potentially split strings, or vector of strings, and join back together, each column becoming 
a single string
}
\author{
m082166
}

