% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.operations.R
\name{NA.operations}
\alias{NA.operations}
\alias{allNA}
\alias{includeNA}
\alias{includeNA.factor}
\alias{includeNA.character}
\title{Some functions to handle NAs}
\usage{
allNA(x)

includeNA(x, label)

\method{includeNA}{factor}(x, label = "(Missing)")

\method{includeNA}{character}(x, label = "(Missing)")
}
\arguments{
\item{x}{An object}

\item{label}{A character string denoting the label to set \code{NA}s to.}
}
\description{
\code{allNA} tests if all elements are NA, and \code{includeNA} sets the
  \code{NA}s in a character vector or factor to an explicit label.
}
\seealso{
\code{\link{is.na}}, \code{\link{anyNA}}
}
\author{
Ethan Heinzen
}
