% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{Nevents}
\alias{Nevents}
\title{Nevents}
\usage{
Nevents(x, ...)
}
\arguments{
\item{x}{a thing}

\item{...}{other arguments}
}
\value{
the events stat from km$table
}
\description{
Number of events in a survival object, within a group
}

