% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{numStats}
\alias{numStats}
\title{numStats}
\usage{
numStats(statistics)
}
\arguments{
\item{statistics}{List holding the data of interest}
}
\value{
Count of elements in statistics.  If that is 1, count of names in that element
}
\description{
Count the number of Elements represented by statistics.  If only one element in statistics, 
count number of elements in that element
}
\author{
m082166
}

