% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.internal.R
\name{paired.internal}
\alias{paired.internal}
\alias{na.paired}
\title{Helper functions for paired}
\usage{
na.paired(object, ...)
}
\arguments{
\item{object}{A \code{data.frame} resulting from evaluating a \code{paired} formula.}

\item{...}{Other arguments, or a vector of indices for extracting.}
}
\value{
\code{na.paired} returns a subsetted version of \code{object} (with attributes).
}
\description{
A set of helper functions for \code{\link{paired}}.
}
\seealso{
\link{tableby.internal}
}
