% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{pasteCenteredStr}
\alias{pasteCenteredStr}
\title{pasteCenteredStr}
\usage{
pasteCenteredStr(strArray, size, sep = " ", padChar = " ")
}
\arguments{
\item{strArray}{Vector of strings to paste together}

\item{size}{Size of each output string. If size < length of any of the strings in strArray, 
strings will be split by makeCenteredStr}

\item{sep}{Separator for paste, defaults to " "}

\item{padChar}{Character to pad out strings to length size, defaults to ' ', 
will cause problems if not nchar 1}
}
\value{
Vector of strings of length size
}
\description{
Make one or more strings consisting of the strings in strArray, separated by sep, 
surrounded by however many instances of padChar are needed to pad out the results 
to centered strings of the requested length.  The strings will be of roughly equal size, 
with a bias for strings to be in decreasing size as we go forward
}
\author{
m082166
}

