% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{pastePaddedStr}
\alias{pastePaddedStr}
\title{pastePaddedStr}
\usage{
pastePaddedStr(strArray, size, sep = " ", padChar = " ",
  appendSep = FALSE)
}
\arguments{
\item{strArray}{Vector of strings to paste together}

\item{size}{Size of each output string. If size < length of any of the strings in strArray, 
strings will be split by makePaddedStr}

\item{sep}{Separator for paste, defaults to " "}

\item{padChar}{Character to pad out strings to length size, defaults to ' ', 
will cause problems if not nchar 1}

\item{appendSep}{If TRUE, and sep and padChar are different, will make sure that n - 1 sep 
appear in output (where n = length (strArray)), defaults to FALSE}
}
\value{
Vector of strings of length size
}
\description{
Make one or more strings consisting of the strings in strArray, separated by sep, 
followed by however many instances of the repeated character are needed to pad out the results 
to strings of the requested length
}
\author{
m082166
}

