% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{process}
\alias{process}
\title{process}
\usage{
process(theText)
}
\arguments{
\item{theText}{Text to parse}
}
\value{
List of strings holding numbers, possibly including a %
}
\description{
Process text, extracting the numbers and returning them as a list of Strings
}
\author{
m082166
}

