% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{sample_size}
\alias{sample_size}
\title{calculate sample size}
\usage{
sample_size(x, Margin.of.Error, conf.level = 0.95, ...)
}
\arguments{
\item{x}{vector}

\item{Margin.of.Error}{numeric}

\item{conf.level}{}
}
\value{
numeric
}
\description{
calculate sample size
}
\examples{



}

