% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{sd2}
\alias{sd2}
\title{Standard Deviation with option for population variance}
\usage{
sd2(x, sample = TRUE)
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
Standard Deviation with option of calculation the population variance
}
\examples{
x <- rnorm(100)
sd2(x, sample = FALSE)

}

