% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{setParam}
\alias{setParam}
\title{setParam}
\usage{
setParam(value, default, testNA = TRUE)
}
\arguments{
\item{value}{The value to test.  If it passes the test, will use it. If not, will use default}

\item{default}{What to use if value isn't valid}

\item{testNA}{If TRUE, reject value if it's NA, if FALSE, reject value if it's NULL.
Defaults to TRUE}
}
\value{
Value, or default
}
\description{
Figure out which parameter value to use, and return it
}
\author{
m082166
}

