% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{showRow}
\alias{showRow}
\title{showRow}
\usage{
showRow(showIntercept, hideRows, rowTitle)
}
\arguments{
\item{showIntercept}{TRUE if should show the Intercept for each line, FALSE if shouldn't}

\item{hideRows}{Vector of names of rows that should not be shown}

\item{rowTitle}{Name of the current row}
}
\value{
TRUE if should show this row, FALSE if shouldn't
}
\description{
Determines whether or not a row should be displayed
}
\author{
m082166
}

