% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{singleSplit}
\alias{singleSplit}
\title{singleSplit}
\usage{
singleSplit(theStr, firstColSize, colSize, strLen)
}
\arguments{
\item{theStr}{String to split}

\item{firstColSize}{The length of the first column}

\item{colSize}{The length of each other column}

\item{strLen}{Total string length}
}
\value{
Contents of each column de-Pandoced but not trimmed
}
\description{
Split a string based on column lengths
}
\author{
m082166
}

