% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{stat_outlier}
\alias{stat_outlier}
\title{stat outlier}
\usage{
stat_outlier(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{stat}
}
\value{
ggplot stat
}
\description{
stat outlier
}
\examples{
mtcars \%>\%
ggplot(aes(x = qsec, y = hp)) +
 geom_point(color = "#112233", size = 3) +
 geom_smooth(method = "lm", se = F, color = "#112233", linetype = 3) +
 stat_outlier(aes(label = rownames(.)), geom = "text", vjust = -1.2, size = 3) +
 xlim(14, 23) +
 ylim(0, 340)


}

