% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\docType{data}
\name{StatOutlier}
\alias{StatOutlier}
\title{get outlir}
\format{An object of class \code{StatChull} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatOutlier
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
get outlier
}
\examples{



}
\keyword{datasets}

