% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.compare.R
\name{summary.compare.data.frame}
\alias{summary.compare.data.frame}
\title{The summary method for a \code{compare.data.frame} object}
\usage{
\method{summary}{compare.data.frame}(object, ..., max.print.vars = NA,
  max.print.obs = NA, max.print.diff = 10, max.print.attrs = NA)
}
\arguments{
\item{object}{An object of class \code{"compare.data.frame"}, as made by the \code{\link{compare.data.frame}} S3 method.}

\item{...}{Other arguments (not in use at this time).}

\item{max.print.vars, max.print.obs, max.print.diff, max.print.attrs}{Integers denoting the maximum number of differences to report
for each of the three tables. Passing \code{NA} will print all differences.}
}
\value{
An object of class \code{"summary.compare.data.frame"} is returned.
}
\description{
Print a more detailed output of the \code{\link{compare.data.frame}} object.
}
\seealso{
compare.data.frame
}
