% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{sumStrLen}
\alias{sumStrLen}
\title{sumStrLen}
\usage{
sumStrLen(strings)
}
\arguments{
\item{strings}{Vector of Strings}
}
\value{
The sum of the nchar lengths of the strings in the Vector
}
\description{
Return the sum of the lengths of the strings in a Vector of strings
}
\author{
m082166
}

