% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{survNinterval}
\alias{survNinterval}
\title{survNinterval}
\usage{
survNinterval(x, x.by, time.interval = 1)
}
\arguments{
\item{x}{a Surv() variable within tableby formula}

\item{x.by}{the by-variable in tableby}

\item{time.interval}{the interval of units of time over which to summarize in categories}
}
\value{
vector of number of events per time interval
}
\description{
survival summary stat per N units of time. Default is years.
}

