% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\docType{data}
\name{theme_flat}
\alias{theme_flat}
\title{theme flat}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
theme_flat
}
\value{
Numeric Index
}
\description{
theme flat
}
\examples{
get_outlier(mtcars$mpg, mtcars$hp)
mtcars[get_outlier(mtcars$mpg, mtcars$hp), ]

}
\keyword{datasets}

