% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{to.data.frame.modelsum}
\alias{to.data.frame.modelsum}
\title{to.data.frame.modelsum}
\usage{
to.data.frame.modelsum(object, results)
}
\arguments{
\item{object}{The data defining the table to display}

\item{results}{List with seven elements:
results: The Pandoc multi-row table
firstColSize: The length of the first column, 
colSize: The length of each other column
showIntercept: TRUE if included the Intercept with each variable
showAdjust: TRUE if included the adjust variable results for each variable
showOnce: Vector of fields to be pulled from element$glance, one time only
translations: List to use for conversion of labels}
}
\value{
A data.frame holding the modelsum
}
\description{
Take a Pandoc Table and turn it into a data frame
}
\author{
m082166
}

