% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsum.R
\name{to.the.data.frame.modelsum}
\alias{to.the.data.frame.modelsum}
\title{to.the.data.frame.modelsum}
\usage{
to.the.data.frame.modelsum(results, header, showOnce, rowCounts, firstColSize,
  colSize, killCol)
}
\arguments{
\item{results}{The Pandoc multi-row table}

\item{header}{Vector with the column headers}

\item{showOnce}{Vector of fields that were pulled from element$glance, so will need to be duplicated}

\item{rowCounts}{List with one element per model of object, holding # rows & y variable}

\item{firstColSize}{The length of the first column,}

\item{colSize}{The length of each other column}

\item{killCol}{Column of results holding data to be ignored}
}
\value{
A data.frame holding the modelsum
}
\description{
Take a Pandoc Table and turn it into a data frame
}
\author{
m082166
}

