% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{trimNumber}
\alias{trimNumber}
\title{trimNumber}
\usage{
trimNumber(number)
}
\arguments{
\item{number}{String to trim}
}
\value{
Resulting String, will not be empty unless number was empty
}
\description{
Take a string representing a number, and remove all trailing 0s, and a trailing "."
}
\author{
m082166
}

