% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{typeTest}
\alias{typeTest}
\title{typeTest}
\usage{
typeTest(outputTypes, whichStat, value)
}
\arguments{
\item{outputTypes}{Vector from which to pull this information. If NULL / NA, will return FALSE}

\item{whichStat}{Which of the items in outputTypes to look at}

\item{value}{(see below)}
}
\value{
TRUE if outputTypes[whichStat] == value output, else FALSE
}
\description{
Returns TRUE if this stat's output type matches value
}
\author{
m082166
}

