% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var2}
\alias{var2}
\title{Variance with option for population variance}
\usage{
var2(x, sample = TRUE)
}
\arguments{
\item{x}{vector or NULL}
}
\value{
none
}
\description{
Variance with option of calculation the population variance
}
\examples{
x <- rnorm(100)
var2(x, sample = FALSE)

}

