% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2.internal.R
\name{write2.internal}
\alias{write2.internal}
\alias{verbatim}
\alias{code.chunk}
\title{Helper functions for \code{write2}}
\usage{
verbatim(...)

code.chunk(..., chunk.opts = "r")
}
\arguments{
\item{...}{For \code{verbatim}, objects to print out monospaced (as if in the terminal). For \code{code.chunk},
either expressions or single character strings to paste into the code chunk.}

\item{chunk.opts}{A single character string giving the code chunk options. Make sure to specify the engine!}
}
\description{
Helper functions for \code{\link{write2}}.
}
\details{
The \code{"verbatim"} class is to tell \code{\link{write2}} to print the object inside
  a section surrounded by three back ticks. The results will look like it would in the terminal (monospaced).

\code{code.chunk()} is to write explicit code chunks in the \code{.Rmd} file; it captures the call and writes it to the
  file, to execute upon knitting.
}
