% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\alias{print.yaml}
\alias{c.yaml}
\alias{is.yaml}
\title{Include a YAML header in \code{write2}}
\usage{
yaml(...)

\method{print}{yaml}(x, ...)

\method{c}{yaml}(..., recursive = FALSE)

is.yaml(x)
}
\arguments{
\item{...}{For \code{yaml()}, arguments to be bundled into a list and passed to \code{\link[yaml]{as.yaml}}.
For \code{print.yaml()}, extra arguments. For \code{c.yaml()}, "yaml" objects to be concatenated.}

\item{x}{An object of class \code{"yaml"}.}

\item{recursive}{Not in use at this time.}
}
\value{
A text string of class \code{"yaml"}.
}
\description{
Include a YAML header in \code{write2}
}
\examples{
x <- yaml(title = "My cool title", author = "Ethan P Heinzen")
x
y <- yaml("header-includes" = list("\\\\usepackage[labelformat=empty]{caption}"))
y
c(x, y)
}
\seealso{
\code{\link[yaml]{as.yaml}}, \code{\link{write2}}
}
\author{
Ethan Heinzen, adapted from an idea by Brendan Broderick
}
