% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-test.R
\name{z.test}
\alias{z.test}
\title{test z test}
\usage{
z.test(x, mu = 0, stdev, alternative = c("two.sided", "less", "greater"),
  sd = stdev, n = length(x), conf.level = 0.95, ...)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
test z test
}
\examples{



}

