% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{arsenal-defunct}
\alias{arsenal-defunct}
\title{Defunct functions in \code{arsenal}}
\arguments{
\item{x, y}{See \code{\link{comparedf}}.}

\item{...}{Other arguments.}
}
\description{
Details about defunct functions in \code{arsenal}
}
\details{
\code{comparison.control} was renamed to \code{\link{comparedf.control}} in version 3.0.0.

\code{compare.data.frame} was renamed to \code{\link{comparedf}} in version 3.0.0.

\code{length.tableby} was removed in version 2.0.0.

\code{includeNA.character} and \code{includeNA.numeric} were removed in version 2.0.0 and
  replaced with a default method.

\code{rangeTime} was removed in version 1.5.0.
}
\seealso{
\code{\link{arsenal-deprecated}}, \code{\link{comparedf}}
}
