% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal-package.R
\docType{package}
\name{arsenal-package}
\alias{arsenal}
\alias{arsenal-package}
\title{arsenal: Useful tools to arm R developers}
\description{
This package is meant to help R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/arsenal}
  \item Report bugs at \url{https://github.com/jimbrig2011/arsenal/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
