% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal.R
\docType{package}
\name{arsenal}
\alias{arsenal}
\alias{arsenal-package}
\title{An arsenal of useful R functions for biomedical statistics}
\description{
An arsenal of useful R functions for biomedical statistics, specializing in models, reporting, and data summaries.
}
\section{Functions}{


Below are listed some of the most widely used functions available in \code{arsenal}:

\code{\link{tableby}}: Summary Statistics Of a Set of Independent Variables By a Categorical Variable.

\code{\link{modelsum}}: Fit models over each of a set of independent variables with a response variable.

\code{\link{freqlist}}: Approximate the output from SAS's PROC FREQ procedure when using the /list option of the TABLE statement.
}

\section{Data}{


\code{\link{mockstudy}}: Mock study data for examples.
}
\examples{
library(arsenal)

}

