% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.freqlist.R
\name{as.data.frame.freqlist}
\alias{as.data.frame.freqlist}
\title{as.data.frame.freqlist}
\usage{
\method{as.data.frame}{freqlist}(x, ..., labelTranslations = NULL,
  list.ok = FALSE)
}
\arguments{
\item{x}{An object of class \code{"freqlist"}.}

\item{...}{Arguments to pass to \code{\link{freq.control}}}

\item{labelTranslations}{A named list (or vector) where the name is the label in the
output to be replaced in the pretty rendering by the character string
value for the named element of the list, e.g., \code{list(age = "Age(Years)", meansd = "Mean(SD)")}.}

\item{list.ok}{If the object has multiple by-variables, is it okay to return a list of data.frames instead of a single data.frame?
If \code{FALSE} but there are multiple by-variables, a warning is issued.}
}
\value{
A data.frame corresponding to the \code{freqlist} object.
}
\description{
Convert \code{\link{freqlist}} object to a data.frame.
}
