% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.internal.R
\name{as.data.frame.freqlist}
\alias{as.data.frame.freqlist}
\title{as.data.frame.freqlist}
\usage{
\method{as.data.frame}{freqlist}(x, row.names, optional, ...)
}
\arguments{
\item{x}{An object of class \code{"freqlist"}.}

\item{row.names}{Not in use at this time, but included for S3 consistency.}

\item{optional}{Not in use at this time, but included for S3 consistency.}

\item{...}{optional arguments included for S3 consistency}
}
\value{
A data.frame corresponding to the \code{freqlist} object.
}
\description{
Convert \code{\link{freqlist}} object to a data.frame.
}

