% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dates}
\alias{dates}
\alias{start_of_month}
\alias{end_of_month}
\title{Start / End of Month Functions}
\usage{
start_of_month(date)

end_of_month(date)
}
\arguments{
\item{date}{Input date (can be character string or a date object)}
}
\value{
Date object containing the Start or End Date of the specified input date's month.
}
\description{
Start / End of Month Functions
}
\examples{
start_of_month(Sys.Date())
end_of_month(Sys.Date())

start_of_month("2019-05-31")
end_of_month("January 15, 2001")
}
