% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.internal.R
\name{freqlist.internal}
\alias{freqlist.internal}
\alias{is.freqlist}
\alias{is.summary.freqlist}
\title{Helper functions for freqlist}
\usage{
is.freqlist(x)

is.summary.freqlist(x)
}
\arguments{
\item{x}{A \code{freqlist} object.}
}
\description{
A set of helper functions for \code{\link{freqlist}}.
}
\seealso{
\code{\link{merge.freqlist}}, \code{\link{arsenal_table}}
}
