% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqlist.internal.R
\name{freqlist.internal}
\alias{freqlist.internal}
\alias{labels<-.freqlist}
\alias{labels.freqlist}
\title{Helper functions for freqlist}
\usage{
\method{labels}{freqlist}(x) <- value

\method{labels}{freqlist}(object, ...)
}
\arguments{
\item{x, object}{A \code{freqlist} object.}

\item{value}{A list or vector of new labels.}

\item{...}{Other arguments (not in use at this time, but included for S3 consistency)}
}
\description{
A set of helper functions for \code{\link{freqlist}}.
}
