% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{geom_mark_outlier}
\alias{geom_mark_outlier}
\title{mark outlier}
\usage{
geom_mark_outlier(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{aes}
}
\value{
ggplot geom
}
\description{
mark outlier in ggplot plot
}
\examples{
mtcars \%>\%
 ggplot(aes(x = mpg, y = hp)) +
 geom_point() +
 geom_smooth(method = "lm", se = F, color = "black", linetype = "dashed") +
 geom_mark_outlier(shape = 1, size = 4, color = "red")


}

