% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{geom_mark_outlier}
\alias{geom_mark_outlier}
\title{geom mark outlier}
\usage{
geom_mark_outlier(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{aes}
}
\value{
ggplot geom
}
\description{
mark outlier in ggplot plot
}
\examples{
mtcars \%>\%
 ggplot(aes(x = qsec, y = hp)) +
 geom_point(color = "#112233", size = 3) +
 geom_smooth(method = "lm", se = F, color = "#112233", linetype = "dotted") +
 geom_mark_outlier(size = 4, alpha = 0.5, color = "#FC575E") +
 xlim(14, 23) +
 ylim(0, 340) +
 theme(panel.background = element_rect(color = "#F0F1F5"))

#'mtcars \%>\%
ggplot(aes(x = qsec, y = hp)) +
 geom_point(color = "#112233", size = 3) +
 geom_smooth(method = "lm", se = F, color = "#112233", linetype = 3) +
 geom_mark_outlier(size = 4, alpha = 0.6, color = "#FC575E") +
 geom_name_outlier(aes(label = rownames(.)), vjust = -1.2, size = 3) +
 xlim(14, 23) +
 ylim(0, 340) +
 theme(panel.background = element_rect(color = "#F0F1F5"))

#'mtcars \%>\%
ggplot(aes(x = factor(cyl), y = hp, label = rownames(.))) +
 geom_boxplot(fill = "#1BA39C", alpha = .6) +
 geom_name_outlier(size = 4, vjust = -1.2) +
 geom_mark_outlier(size = 3, color = "#FC575E", alpha = 0.7) +
 xlab("cyl") +
 theme(panel.background = element_rect(color = "#F0F1F5"))

}

