% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{geom_name_outlier}
\alias{geom_name_outlier}
\title{name outlier}
\usage{
geom_name_outlier(mapping = NULL, data = NULL,
  position = position_stack(1.07), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{aes}
}
\value{
ggplot geom
}
\description{
name outlier in ggplot plot
}
\examples{
mtcars \%>\%
ggplot(aes(x = factor(cyl), y = hp, label = rownames(.))) +
 geom_boxplot() +
 geom_name_outlier(size = 4)


}

