% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R
\name{internal.functions}
\alias{internal.functions}
\alias{smart.split}
\title{Split a string into pieces intelligently}
\usage{
smart.split(string, width = Inf, min.split = -Inf)
}
\arguments{
\item{string}{A character vector}

\item{width}{Either \code{Inf} or \code{NULL} to specify no splitting,
or a positive integer giving the largest allowed string length.}

\item{min.split}{Either \code{-Inf} or \code{NULL} to specify no
lower bound on the string length, or a positive integer giving the minimum string length.}
}
\value{
A list of the same length as \code{string}, with each element being
  the "intelligently" split string.
}
\description{
Split a string into pieces intelligently
}
