% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.labels.R
\name{keep.labels}
\alias{keep.labels}
\alias{keep.labels}
\alias{keep.labels.data.frame}
\alias{keep.labels.default}
\alias{[.keep_labels}
\title{Keep Labels}
\usage{
keep.labels(x, ...)

\method{keep.labels}{data.frame}(x, ...)

\method{keep.labels}{default}(x, ...)

\method{[}{keep_labels}(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{Other arguments (not in use at this time).}
}
\value{
A copy of \code{x} with a "keep labels" class appended on.
}
\description{
Keep the \code{'label'} attribute on an R object when subsetting.
}
\seealso{
\code{\link{labels}}
}
\author{
Ethan Heinzen
}
