% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{List files excluding temporary, Rproj, and git removing internal
undesired files}
\usage{
list_files(path, pat = NULL, full = FALSE, named = FALSE)
}
\arguments{
\item{path}{(String) Valid path}

\item{pat}{(String) Pattern to match. Defaults to NULL.}

\item{full}{(Logical) Return full path? Defaults to FALSE.}

\item{named}{(Logical) Add names? Defaults to FALSE.}
}
\value{
A character vector containing the files in the specified directory
excluding any temporary unwanted files (i.e. "~$file.xlsx"), Rproj files, and git files.
}
\description{
List files excluding temporary, Rproj, and git removing internal
undesired files
}
\examples{
list_files(getwd())
}
