% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.functions.R, R/modelsum.internal.R
\name{[.modelsum}
\alias{[.modelsum}
\alias{labels.modelsum}
\alias{labels<-.modelsum}
\alias{merge.modelsum}
\alias{modelsum.internal}
\alias{is.modelsum}
\alias{is.summary.modelsum}
\alias{na.modelsum}
\title{Helper functions for modelsum}
\usage{
\method{[}{modelsum}(x, i, j, ..., warn = TRUE)

\method{labels}{modelsum}(object, ...)

\method{labels}{modelsum}(x) <- value

\method{merge}{modelsum}(x, y, all = FALSE, all.x = all, all.y = all,
  ...)

is.modelsum(x)

is.summary.modelsum(x)

na.modelsum(object, ...)
}
\arguments{
\item{x, y}{A \code{modelsum} object.}

\item{i}{A vector to index \code{x} with: either names of variables, a numeric vector, or a logical vector of appropriate length.
\code{i} indexes the x-variables, and \code{j} indexes the by-variables.}

\item{j}{A vector to index \code{x} with: either names of variables, a numeric vector, or a logical vector of appropriate length.
\code{i} indexes the x-variables, and \code{j} indexes the by-variables.}

\item{...}{Other arguments, or a vector of indices for extracting.}

\item{warn}{A logical, indicating whether to warn if some indices aren't found.}

\item{object}{A \code{data.frame} resulting from evaluating a \code{modelsum} formula.}

\item{value}{A list of new labels.}

\item{all}{Logicals, denoting which terms to keep if not all are in common.}

\item{all.x}{Logicals, denoting which terms to keep if not all are in common.}

\item{all.y}{Logicals, denoting which terms to keep if not all are in common.}
}
\value{
\code{na.modelsum} returns a subsetted version of \code{object} (with attributes).
}
\description{
A set of helper functions for \code{\link{modelsum}}.
}
