% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsum.internal.R
\name{modelsum.internal}
\alias{modelsum.internal}
\alias{na.modelsum}
\alias{[.modelsum}
\alias{labels.modelsum}
\alias{labels<-.modelsum}
\alias{merge.modelsum}
\title{Helper functions for modelsum}
\usage{
na.modelsum(object, ...)

\method{[}{modelsum}(x, i)

\method{labels}{modelsum}(object, ...)

\method{labels}{modelsum}(x) <- value

\method{merge}{modelsum}(x, y, ...)
}
\arguments{
\item{object}{A \code{data.frame} resulting from evaluating a \code{modelsum} formula.}

\item{...}{Other arguments, or a vector of indices for extracting.}

\item{x, y}{A \code{modelsum} object.}

\item{i}{A vector to index \code{x} with: either names of variables, a numeric vector, or a logical vector of appropriate length.}

\item{value}{A list of new labels.}
}
\value{
\code{na.modelsum} returns a subsetted version of \code{object} (with attributes).
}
\description{
A set of helper functions for \code{\link{modelsum}}.
}
