% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsum.internal.R
\name{modelsum.internal}
\alias{modelsum.internal}
\alias{is.modelsum}
\alias{is.summary.modelsum}
\alias{na.modelsum}
\title{Helper functions for modelsum}
\usage{
is.modelsum(x)

is.summary.modelsum(x)

na.modelsum(object, ...)
}
\arguments{
\item{x}{A \code{modelsum} object.}

\item{object}{A \code{data.frame} resulting from evaluating a \code{modelsum} formula.}

\item{...}{Other arguments, or a vector of indices for extracting.}
}
\value{
\code{na.modelsum} returns a subsetted version of \code{object} (with attributes).
}
\description{
A set of helper functions for \code{\link{modelsum}}.
}
\seealso{
\code{\link{arsenal_table}}
}
