% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arsenal.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compare}
\title{Objects exported from other packages}
\seealso{
\code{\link{compare.data.frame}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{testthat}{\code{\link[testthat]{compare}}}
}}

