% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-outlier.R
\name{stat_outlier}
\alias{stat_outlier}
\title{stat outlier}
\usage{
stat_outlier(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{x}{stat}
}
\value{
ggplot stat
}
\description{
stat outlier
}
\examples{



}

