% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.comparedf.R
\name{summary.compare}
\alias{summary.compare}
\alias{summary.comparedf}
\alias{print.summary.comparedf}
\title{The summary method for a \code{comparedf} object}
\usage{
\method{summary}{comparedf}(object, ..., show.attrs = FALSE)

\method{print}{summary.comparedf}(x, ..., format = "pandoc")
}
\arguments{
\item{object}{An object of class \code{"comparedf"}, as made by the \code{\link{comparedf}} S3 method.}

\item{...}{Other arguments passed to \code{\link{comparedf.control}}. In \code{print}, these are passed to \code{\link[knitr]{kable}}.}

\item{show.attrs}{Logical, denoting whether to show the actual attributes which are different. For (e.g.) factors with lots
of levels, this can make the tables quite wide, so this feature is \code{FALSE} by default.}

\item{x}{An object returned by the \code{summary.comparedf} function.}

\item{format}{Passed to \code{\link[knitr]{kable}}: the format for the table. The default here is "pandoc".
To use the default in \code{kable}, pass \code{NULL}.}
}
\value{
An object of class \code{"summary.comparedf"} is returned.
}
\description{
Print a more detailed output of the \code{\link{comparedf}} object.
}
\seealso{
\code{\link{comparedf}}, \code{\link{comparedf.control}}
}
