% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{tableby.internal}
\alias{tableby.internal}
\alias{merge.tableby}
\alias{modpval.tableby}
\alias{labels.tableby}
\alias{labels<-}
\alias{tests}
\alias{tests.tableby}
\alias{labels<-.tableby}
\alias{[.tableby}
\alias{na.tableby}
\title{Helper functions for tableby}
\usage{
\method{merge}{tableby}(x, y, ...)

modpval.tableby(x, pdata, use.pname = FALSE)

\method{labels}{tableby}(object, ...)

labels(x) <- value

tests(x)

\method{tests}{tableby}(x)

\method{labels}{tableby}(x) <- value

\method{[}{tableby}(x, ...)

na.tableby(object, ...)
}
\arguments{
\item{x, y}{A \code{tableby} object.}

\item{...}{Other arguments, or a vector of indices for extracting.}

\item{pdata}{A named data.frame where the first column is the x variable names matched by name, the second is the
p-values (or some test stat), and the third column is the method name (optional)}

\item{use.pname}{Logical, denoting whether the column name in \code{pdata} corresponding to the p-values should be used
in the output of the object.}

\item{object}{A \code{data.frame} resulting form evaluating \code{modelsum} formula.}

\item{value}{A list of new labels.}
}
\value{
\code{na.tableby} returns a subsetted version of \code{object} (with attributes).
}
\description{
A set of helper functions for \code{\link{tableby}}.
}
