% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby.internal.R
\name{tableby.stats.internal}
\alias{tableby.stats.internal}
\alias{as.tbstat}
\alias{as.countpct}
\alias{as.tbstat_multirow}
\title{Internal \code{tableby} functions}
\usage{
as.tbstat(x, oldClass = NULL, sep = NULL, parens = NULL, sep2 = NULL,
  pct = NULL)

as.countpct(x, ...)

as.tbstat_multirow(x)
}
\arguments{
\item{x}{Usually a vector.}

\item{oldClass}{class(es) to add to the resulting object.}

\item{sep}{The separator between \code{x[1]} and the rest of the vector.}

\item{parens}{A length-2 vector denoting parentheses to use around \code{x[2]} and \code{x[3]}.}

\item{sep2}{The separator between \code{x[2]} and \code{x[3]}.}

\item{pct}{The symbol to use after percents.}

\item{...}{arguments to pass to \code{as.tbstat}.}
}
\description{
A collection of functions that may help users create custom functions that are formatted correctly.
}
\details{
\code{as.tbstat} defines a tableby statistic with its appropriate formatting.

  \code{as.countpct} adds another class to \code{as.tbstat} to use different "digits" arguments. See \code{\link{tableby.control}}.

  \code{as.tbstat_multirow} marks an object (usually a list) for multiple-row printing.
}
