% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2.internal.R
\name{write2.internal}
\alias{print.verbatim}
\alias{verbatim}
\alias{write2.internal}
\title{Helper functions for \code{write2}}
\usage{
\method{print}{verbatim}(x, ...)

verbatim(x)
}
\arguments{
\item{x}{An R object to coerce to class \code{"verbatim"} or such an object to print.}

\item{...}{Other arguments passed to \code{print}.}
}
\description{
Helper functions for \code{\link{write2}}.
}
\details{
The \code{"verbatim"} class is to tell \code{\link{write2}} to print the object inside
  a section surrounded by three back ticks. The results will look like it would in the terminal (monospaced).
}

