% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2.internal.R
\name{write2.internal}
\alias{as.threeticks}
\alias{print.threeticks}
\alias{write2.internal}
\title{Helper functions for \code{write2}}
\usage{
\method{print}{threeticks}(x, ...)

as.threeticks(x)
}
\arguments{
\item{x}{An R object to coerce to class \code{"threeticks"} or such an object to print.}

\item{...}{Other arguments passed to \code{print}.}
}
\description{
Helper functions for \code{\link{write2}}.
}

